+++ Client/src/Pages/Home/Home.jsx
@@ -1,15 +1,16 @@
-import React from 'react';
-import Banner from './Banner/Banner';
-import FetureTopSection from './FetureTopSection/FetureTopSection';
+import React from "react";
+import Banner from "./Banner/Banner";
+import FetureTopSection from "./FetureTopSection/FetureTopSection";
+import Purchase from "./Purchase/Purchase";
 
 const Home = () => {
-    return (
-        <div>
-            <Banner />
-            <FetureTopSection />
-           
-        </div>
-    );
+  return (
+    <div>
+      <Banner />
+      <FetureTopSection />
+      <Purchase />
+    </div>
+  );
 };
 
 export default Home;
