--- Client/src/Pages/Share/Footer/Footer.jsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-const Footer = () => {
-  return (
-    <div>
-      <p></p>
-    </div>
-  );
-};
-
-export default Footer;
