+++ Client/src/Share/Footer/Footer.jsx
@@ -3,7 +3,7 @@ import React from 'react';
 const Footer = () => {
     return (
         <div>
-            <h1 className='text-6xl text-center text-red-700'>this is out footer</h1>
+           <p></p> 
         </div>
     );
 };
